only run cleanup if there was an update

+++ src/storage/local.js
@@ -27,7 +27,7 @@ export default class LocalStorage extends BackupStorage {
   /**
    * Override of the default method
    */
-  async backupAsset(asset) {
+  async backupAsset({ asset, existing }) {
     if (!fs.existsSync(this.getAssetDirectory(asset))) {
       fs.mkdirSync(this.getAssetDirectory(asset), { recursive: true })
     }
@@ -35,9 +35,11 @@ export default class LocalStorage extends BackupStorage {
     try {
       fs.writeFileSync(metaDataPath, JSON.stringify(asset, null, 4))
       await this.downloadAsset(asset)
-      const metadataFiles = glob.sync(`${this.getAssetDirectory(asset)}/sb_asset_data_*.json`).filter(f => f !== metaDataPath)
-      for (const metadataFile of metadataFiles) {
-        fs.rmSync(metadataFile)
+      if (existing) {
+        const metadataFiles = glob.sync(`${this.getAssetDirectory(asset)}/sb_asset_data_*.json`).filter(f => f !== metaDataPath)
+        for (const metadataFile of metadataFiles) {
+          fs.rmSync(metadataFile)
+        }
       }
       return true
     } catch (err) {
