+++ src/storage/s3.js
@@ -31,7 +31,7 @@ export default class S3Storage extends BackupStorage {
   /**
    * Override of the default method
    */
-  async backupAsset(asset) {
+  async backupAsset({ asset, existing }) {
     if (!fs.existsSync(this.getAssetDirectory(asset))) {
       fs.mkdirSync(this.getAssetDirectory(asset), { recursive: true })
     }
@@ -42,9 +42,11 @@ export default class S3Storage extends BackupStorage {
       const assetStream = fs.createReadStream(`${this.getAssetDirectory(asset)}/${filename}`)
       await this.s3Client.putObject({ Bucket: this.bucket, Key: `${this.spaceId}/${asset.id}/${filename}`, Body: assetStream }).promise()
       fs.rmdirSync(this.getAssetDirectory(asset), { recursive: true })
-      const r = await this.s3Client.listObjectsV2({Bucket: `${this.bucket}`, Prefix: `${this.spaceId}/${asset.id}/sb_asset_data_`}).promise()
-      const metadataFiles = r.Contents.filter(f => f.Key !== `${this.spaceId}/${asset.id}/${this.getAssetDataFilename(asset)}`)
-      await this.s3Client.deleteObjects({Bucket: `${this.bucket}`, Delete: {Objects: metadataFiles.map(f => {return {Key: f.Key}})}}).promise()
+      if (existing) {
+        const r = await this.s3Client.listObjectsV2({ Bucket: `${this.bucket}`, Prefix: `${this.spaceId}/${asset.id}/sb_asset_data_` }).promise()
+        const metadataFiles = r.Contents.filter(f => f.Key !== `${this.spaceId}/${asset.id}/${this.getAssetDataFilename(asset)}`)
+        await this.s3Client.deleteObjects({ Bucket: `${this.bucket}`, Delete: { Objects: metadataFiles.map(f => { return { Key: f.Key } }) } }).promise()
+      }
       return true
     } catch (err) {
       console.error(err)
